+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2017-11-09 - SC
+### Added
+- phpunit / travis-ci
+- trivial test to see if pipeline works
+
 ## [Unreleased] - SC
 ### Changed
 - Flight view now uses community names